requirements.txt
aiogram==3.4.1
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
alembic==1.13.1
python-dotenv==1.0.0
aiofiles==23.2.1
sentry-sdk==1.40.0
